Refactor termination analysis
